backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.7
mysql-connector-python==8.2.0
pydantic==2.5.0
python-multipart==0.0.6
python-cors==1.0.0
jinja2==3.1.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
email-validator==2.1.0
asyncio-mqtt==0.13.0
python-dotenv==1.0.0
